Moved UiWindow class declaration to its own cc file. Move dialogs there too becuase they're small and the only ones to use UiWindow
